Fix level and type info
